pyproject.toml
[tool.poetry]
name = "django-gunicorn-nginx-docker"
version = "1.0.0"
description = ""
authors = ["Watt Iamsuri <wiamsuri@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
Django = "^4.2"
gunicorn = "^20.1.0"
djangorestframework = "^3.15.1"
djangorestframework-camel-case = "^1.4.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
django-storages = "^1.14.3"
django-filter = "^24.2"
pillow = "^10.3.0"
django-environ = "^0.11.2"
boto3 = "^1.34.108"
channels = "^4.1.0"
daphne = "^4.1.2"
channels-redis = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
